+++ miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/vo/MyVoucherVo.java
@@ -27,8 +27,8 @@ public class MyVoucherVo {
     @Column(commit = "代金券金额")
     private Double price;
 
-    @Column(commit = "满减下限", length = 11)
-    private Integer reductionLimit;
+    @Column(commit = "满减下限", length = 11,precision = 2)
+    private Double reductionLimit;
 
     @Column(commit = "预设有效时长(时间戳)")
     private Long effectiveTime;
